feat: Create window via content script relay directly
1) Remove the passage of window.addEventListener() and
window.postMessage() in the content script.
2) Use the content script relay directly to have the service
worker create a window.
